Address phpstan issues

+++ includes/main.php
@@ -15,12 +15,12 @@ add_action( 'template_redirect', __NAMESPACE__ . '\handle_request', 1 );
 /**
  * Enqueue the script used by this plugin.
  */
-function enqueue_scripts() {
+function enqueue_scripts(): void {
 	wp_enqueue_script(
 		'quick-escape-widget',
 		plugin_dir_url( __DIR__ ) . '/build/index.js',
 		[],
-		filemtime( plugin_dir_path( __DIR__ ) . 'build/index.js' ),
+		(string) filemtime( plugin_dir_path( __DIR__ ) . 'build/index.js' ),
 		true
 	);
 }
@@ -28,7 +28,7 @@ function enqueue_scripts() {
 /**
  * Register the widget provided by this plugin.
  */
-function register_widget() {
+function register_widget(): void {
 	\register_widget( 'QuickEscapeWidget\Widget\Quick_Escape_Widget' );
 }
 
@@ -36,10 +36,10 @@ function register_widget() {
  * Move a stored-as-transient redirect URL into its more permanent position
  * if one is found.
  */
-function check_redirect_storage() {
+function check_redirect_storage(): void {
 	$redirect_url = get_option( '_transient_qew_redirect_url', '' );
 
-	if ( '' !== $redirect_url ) {
+	if ( is_string( $redirect_url ) && wp_http_validate_url( $redirect_url ) ) {
 		update_option( 'qew_redirect_url', esc_url_raw( $redirect_url ) );
 		delete_option( '_transient_qew_redirect_url' );
 	}
@@ -49,13 +49,17 @@ function check_redirect_storage() {
  * Handle a redirect request registered in the browser's
  * history by a click on the Quick Escape link.
  */
-function handle_request() {
+function handle_request(): void {
 	if ( isset( $_REQUEST['qew'] ) && 1 === (int) $_REQUEST['qew'] ) { // phpcs:ignore WordPress.Security.NonceVerification.Recommended
 		$redirect_url = get_option( 'qew_redirect_url', '' );
 
 		// The old transient data may still be in place.
 		if ( '' === $redirect_url ) {
 			$redirect_url = get_option( '_transient_qew_redirect_url', 'https://google.com' );
+		}
+
+		if ( ! is_string( $redirect_url ) || ! wp_http_validate_url( $redirect_url ) ) {
+			$redirect_url = 'https://google.com';
 		}
 
 		wp_redirect( $redirect_url ); // phpcs:ignore WordPress.Security.SafeRedirect.wp_redirect_wp_redirect
